外部flash调通，发现存储0xFF数据出错

+++ Drivers/BSP/w25q64.h
@@ -32,7 +32,7 @@ typedef __I int32_t vsc32;  /*!< Read Only */
 typedef __I int16_t vsc16;  /*!< Read Only */
 typedef __I int8_t vsc8;   /*!< Read Only */
 
-typedef uint32_t  u32;
+typedef unsigned long  u32;
 typedef uint16_t u16;
 typedef uint8_t  u8;
